Adapt for python 3

+++ include/utils/train_utils.py
@@ -260,10 +260,10 @@ def interp(w, i, channel_dim):
         of the same length == len(i)
     '''
     w_as_vector = tf.reshape(w, [-1, channel_dim]) # gather expects w to be 1-d
-    upper_l = tf.to_int32(tf.concat(axis=1, values=[i[:, 0:1], tf.floor(i[:, 1:2]), tf.floor(i[:, 2:3])]))
-    upper_r = tf.to_int32(tf.concat(axis=1, values=[i[:, 0:1], tf.floor(i[:, 1:2]), tf.ceil(i[:, 2:3])]))
-    lower_l = tf.to_int32(tf.concat(axis=1, values=[i[:, 0:1], tf.ceil(i[:, 1:2]), tf.floor(i[:, 2:3])]))
-    lower_r = tf.to_int32(tf.concat(axis=1, values=[i[:, 0:1], tf.ceil(i[:, 1:2]), tf.ceil(i[:, 2:3])]))
+    upper_l = tf.cast(tf.concat(axis=1, values=[i[:, 0:1], tf.floor(i[:, 1:2]), tf.floor(i[:, 2:3])]), tf.int32)
+    upper_r = tf.cast(tf.concat(axis=1, values=[i[:, 0:1], tf.floor(i[:, 1:2]), tf.ceil(i[:, 2:3])]), tf.int32)
+    lower_l = tf.cast(tf.concat(axis=1, values=[i[:, 0:1], tf.ceil(i[:, 1:2]), tf.floor(i[:, 2:3])]), tf.int32)
+    lower_r = tf.cast(tf.concat(axis=1, values=[i[:, 0:1], tf.ceil(i[:, 1:2]), tf.ceil(i[:, 2:3])]), tf.int32)
 
     upper_l_idx = to_idx(upper_l, tf.shape(w))
     upper_r_idx = to_idx(upper_r, tf.shape(w))
@@ -320,5 +320,5 @@ def bilinear_select(H, pred_boxes, early_feat, early_feat_channels, w_offset, h_
                                           0,
                                           scale_factor * H['grid_height'] - 1)
 
-    interp_indices = tf.concat(axis=1, values=[tf.to_float(batch_ids), pred_y_center_clip, pred_x_center_clip])
+    interp_indices = tf.concat(axis=1, values=[tf.cast(batch_ids, tf.float32), pred_y_center_clip, pred_x_center_clip])
     return interp_indices
